% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calc_dxy}
\alias{calc_dxy}
\title{Function to calculate the average number of nucleotide differences (dxy) between two separate populations (wakeley)}
\usage{
calc_dxy(l)
}
\arguments{
\item{l}{List with objects of class DNAbin or matrices where rows are sequences and columns are nucleotide positions from two different populations}
}
\description{
Function to calculate the average number of nucleotide differences (dxy) between two separate populations (wakeley)
}
