% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calc_pi}
\alias{calc_pi}
\title{Calculate the average number of nucleotide differences (Tajima 1983 eq. A3)}
\usage{
calc_pi(x)
}
\arguments{
\item{x}{An object of class DNAbin or a matrix where rows are sequences and columns are nucleotide positions (assumed to be already aligned)}
}
\description{
Calculate the average number of nucleotide differences (Tajima 1983 eq. A3)
}
