% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calc_pi_pops}
\alias{calc_pi_pops}
\title{Function to calculate the average number of nucleotide differences between two separate populations}
\usage{
calc_pi_pops(x, y)
}
\arguments{
\item{x, y}{Objects of class DNAbin or matrices where rows are sequences and columns are nucleotide positions from two different populations}
}
\description{
Function to calculate the average number of nucleotide differences between two separate populations
}
