% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{concat_seqs}
\alias{concat_seqs}
\title{Concatenate sequences into a vector from matrix}
\usage{
concat_seqs(x)
}
\arguments{
\item{x}{A matrix of a multiple sequence alignment where each column is a base
pair and each row is a sequence}
}
\description{
Concatenate sequences into a vector from matrix
}
