% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{count_tstv}
\alias{count_tstv}
\title{Function to calculate simple transition/transversion ratio and transition rate}
\usage{
count_tstv(x)
}
\arguments{
\item{x}{Object of class DNAbin}
}
\description{
Function to calculate simple transition/transversion ratio and transition rate
}
