% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read.arp}
\alias{read.arp}
\title{Read in arp files of single DNA Loci}
\usage{
read.arp(file)
}
\arguments{
\item{file}{PATH to .arp file}
}
\description{
This function reads in .arp files from fsc26 and creates DNAbin
objects for each simulated population
}
