% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sharedhaps}
\alias{sharedhaps}
\title{Function to calculate the number of shared haplotypes between populations}
\usage{
sharedhaps(x)
}
\arguments{
\item{x}{A list of matrices of multiple sequence alignments or DNAbins}
}
\description{
Function to calculate the number of shared haplotypes between populations
}
