% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sumstats}
\alias{sumstats}
\title{Function to calculate the number of haplotypes, number of segregating sites, haplotype diversity, and pi of a matrix of sequences}
\usage{
sumstats(x)
}
\arguments{
\item{x}{Object of class DNAbin or a matrix of multiple sequence alignment where
rows are sequences and columns are nucleotide positions}
}
\description{
Function to calculate the number of haplotypes, number of segregating sites, haplotype diversity, and pi of a matrix of sequences
}
