% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tajimaD}
\alias{tajimaD}
\title{Function to calculate Tajima's D}
\usage{
tajimaD(x)
}
\arguments{
\item{x}{Object of class DNAbin or a matrix of multiple sequence alignment where
rows are sequences and columns are nucleotide positions}
}
\description{
Function to calculate Tajima's D
}
