% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{theta_watt}
\alias{theta_watt}
\title{Function to calculate the Watterson estimator for theta}
\usage{
theta_watt(x)
}
\arguments{
\item{x}{Object of class DNAbin or a matrix of multiple sequence alignment where
rows are sequences and columns are nucleotide positions}
}
\description{
Function to calculate the Watterson estimator for theta
}
