% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sumstats_pops}
\alias{sumstats_pops}
\title{Function to calculate sumstats on samples with multiple populations}
\usage{
sumstats_pops(x)
}
\arguments{
\item{x}{List of objects of class DNAbin or matrices of multiple sequence alignment where
rows are sequences and columns are nucleotide positions. Each object in list is a separate
population.}
}
\description{
Function to calculate sumstats on samples with multiple populations
}
